% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\docType{package}
\name{BLBLogistic-package}
\alias{BLBLogistic}
\alias{BLBLogistic-package}
\title{BLBLogistic: Bag of Little Bootstrap for Logistic Regression Model}
\description{
LogisticRegressionBLB uses bag of little bootstraps to estimate regression coefficients, sigma, and prediction intervals on a binary logit model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipalvis/BLBLogistic}
  \item Report bugs at \url{https://github.com/ipalvis/BLBLogistic/issues}
}

}
\author{
\strong{Maintainer}: Alvis Ip \email{aip@ucdavis.edu}

}
