% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{estimate the regression estimates based on given number of repetitions}
\usage{
glm1(formula, data, freqs)
}
\description{
estimate the regression estimates based on given number of repetitions
}
