% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{This computes estimates for the partitioned subsamples from split_data.}
\usage{
glm_each_subsample(formula, data, n, B)
}
\description{
This computes estimates for the partitioned subsamples from split_data.
}
