% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\docType{package}
\name{LogisticRegressionBLB-package}
\alias{LogisticRegressionBLB}
\alias{LogisticRegressionBLB-package}
\title{LogisticRegressionBLB: Bag of Little Bootstrap for Logistic Regression Model}
\description{
LogisticRegressionBLB uses bag of little bootstraps to estimate regression coefficients, sigma, and prediction intervals on a binary logit model.
}
\author{
\strong{Maintainer}: Alvis Ip \email{aip@ucdavis.edu}

}
