% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blbglm.R
\name{glm1}
\alias{glm1}
\title{Based on freqs - the number of repetitions, this generates the glm model for each and determines the coefficients and sigmas of each BLB.}
\usage{
glm1(formula, data, freqs)
}
\description{
Based on freqs - the number of repetitions, this generates the glm model for each and determines the coefficients and sigmas of each BLB.
}
